refactor(slima2a,srpc): Enable some addition ruff rules and fix issues
